chore: Исправления после ревью
Разбила view на более мелкие. Добавила preview. Улучшила структуру файлов.
