+++ travel schedule/Mock/Extension+CarrierInfo.swift
@@ -1,0 +1,18 @@
+//
+//  Extension+CarrierInfo.swift
+//  travel schedule
+//
+//  Created by user on 28.06.2025.
+//
+
+import Foundation
+
+// MARK: - Mock Data for CarrierInfo
+extension CarrierInfo {
+    static let mock = CarrierInfo(
+        title: "Аэрофлот",
+        email: "callcenter@aeroflot.ru",
+        phoneNumber: "+7 (495) 223-55-55",
+        imageURL: URL(string: "https://yastat.net/s3/rasp/media/data/company/logo/aeroflot_logo_ru.gif")
+    )
+}
