+++ travel schedule/Views/MainView/Model/DeparturePoint.swift
@@ -1,0 +1,13 @@
+//
+//  DeparturePoint.swift
+//  travel schedule
+//
+//  Created by user on 28.06.2025.
+//
+
+import SwiftUI
+
+struct DeparturePoint {
+    var city: String
+    var station: String
+}
