+++ travel schedule/Views/MainView/Model/Navigation/Destination.swift
@@ -1,0 +1,18 @@
+//
+//  Destination.swift
+//  travel schedule
+//
+//  Created by user on 28.06.2025.
+//
+
+import Foundation
+
+enum Destination: Hashable {
+    case stationFrom
+    case stationTo
+    case from
+    case to
+    case listOfCarriers
+    case carrierInfo
+    case filter
+}
