+++ travel schedule/Views/SelectionView/StationSelectionView.swift
@@ -4,7 +4,7 @@ struct StationSelectionView: View {
     private let iconSize: Double = 24
     private let rowHeight: Double = 60
     
-    @StateObject var viewModel = MockViewModel()
+    @StateObject var viewModel = SelectionDataViewModelMock()
     @Binding var selectedStation: String
     @Binding var selectedCity: String
     @Binding var path: NavigationPath
