+++ travel schedule/Views/ServerErrorView.swift
@@ -3,19 +3,21 @@ import SwiftUI
 enum ErrorViewType: String {
     case serverError
     case networkError
+    
+    var image: Image {
+        Image(self.rawValue)
+    }
+    
+    var title: String {
+        switch self {
+        case .networkError: return "Нет интернета"
+        case .serverError: return "Ошибка сервера"
+        }
+    }
 }
 
 struct ErrorView: View {
     let errorType: ErrorViewType
-    let image: Image
-    let text: String
-    
-    init(errorType: ErrorViewType) {
-        self.errorType = errorType
-        self.image = Image(errorType.rawValue)
-        self.text = errorType == .networkError ? "Нет интернета" : "Ошибка сервера"
-    }
-    
     private let imageSize: Double = 223
     
     var body: some View {
@@ -23,12 +25,12 @@ struct ErrorView: View {
             Color.customWhite
                 .ignoresSafeArea()
             VStack(spacing: 16) {
-                image
+                errorType.image
                     .resizable()
                     .aspectRatio(contentMode: .fit)
                     .frame(width: imageSize, height: imageSize)
                     .clipShape(RoundedRectangle(cornerRadius: 70))
-                Text(text)
+                Text(errorType.title)
                     .font(.system(size: 24, weight: .bold))
                     .foregroundStyle(Color.customBlack)
             }
@@ -39,4 +41,5 @@ struct ErrorView: View {
 
 #Preview {
     ErrorView(errorType: .serverError)
+    ErrorView(errorType: .networkError)
 }
