+++ travel schedule/Views/StoriesView/ProgressBar/View/MaskFragmetView.swift
@@ -1,0 +1,17 @@
+//
+//  MaskFragmetView.swift
+//  travel schedule
+//
+//  Created by user on 28.06.2025.
+//
+
+import SwiftUI
+
+struct MaskFragmentView: View {
+    var body: some View {
+        RoundedRectangle(cornerRadius: .progressBarCornerRadius)
+            .fixedSize(horizontal: false, vertical: true)
+            .frame(height: .progressBarHeight)
+            .foregroundStyle(.white)
+    }
+}
