+++ travel schedule/Views/StoriesView/StoriesTabView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct StoriesTabView: View {
-    let stories: [StoriesScreen]
+    let stories: [StoryScreen]
     @Binding var currentStoryIndex: Int
 
     var body: some View {
@@ -16,9 +16,6 @@ struct StoriesTabView: View {
             TabView(selection: $currentStoryIndex) {
                 ForEach(Array(stories.enumerated()), id: \.element.id) { index, story in
                     StoryView(story: story)
-//                        .onTapGesture {
-//                            didTapStory()
-//                        }
                         .tag(index)
                 }
             }
@@ -39,7 +36,31 @@ struct StoriesTabView: View {
     func previousStoryTap() {
         currentStoryIndex = max(currentStoryIndex - 1, 0)
     }
-//    func didTapStory() {
-//        currentStoryIndex = min(currentStoryIndex + 1, stories.count - 1)
-//    }
+}
+
+
+#Preview {
+    struct StoriesTabViewPreviewWrapper: View {
+        @State private var currentStoryIndex = 0
+        
+        var body: some View {
+            StoriesTabView(
+                stories: [
+                    StoryScreen(
+                        title: "Story 1",
+                        description: "Description 1",
+                        image: Image("firstImage1")
+                    ),
+                    StoryScreen(
+                        title: "Story 2",
+                        description: "Description 2",
+                        image: Image("firstImage2")
+                    )
+                ],
+                currentStoryIndex: $currentStoryIndex
+            )
+        }
+    }
+    
+    return StoriesTabViewPreviewWrapper()
 }
